+++ src/dropdown/dropdown-control.directive.ts
@@ -1,0 +1,9 @@
+import { Directive } from '@angular/core';
+
+@Directive({
+    selector: '[fd-dropdown-control]',
+    host: {
+        class: 'fd-dropdown__control'
+    }
+})
+export class DropdownControlDirective {}
