+++ src/dropdown/dropdown-group.component.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'fd-dropdown-group',
+    host: {
+        class: 'fd-dropdown__group'
+    },
+    template: `
+    <ng-content></ng-content>
+  `
+})
+export class DropdownGroupComponent {}
