+++ src/dropdown/dropdown-header.directive.ts
@@ -1,0 +1,10 @@
+import { Directive } from '@angular/core';
+
+@Directive({
+    selector: 'fd-dropdown-header',
+    host: {
+        class: 'fd-dropdown__title',
+        role: 'separator'
+    }
+})
+export class DropdownGroupHeaderDirective {}
